# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_pranavbh-9117_IMB_internal_institution_dto.CreateRequest:
    properties:
      address:
        type: string
      code:
        maxLength: 50
        type: string
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - code
    - name
    type: object
  github_com_pranavbh-9117_IMB_internal_institution_dto.InstitutionResponse:
    properties:
      address:
        type: string
      code:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
  github_com_pranavbh-9117_IMB_internal_institution_dto.UpdateRequest:
    properties:
      address:
        type: string
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 20
        type: string
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse:
    properties:
      data: {}
      message:
        example: error description
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any:
    properties:
      data: {}
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  ? github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_institution_dto.InstitutionResponse'
        type: array
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  ? github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_institution_dto.InstitutionResponse'
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the Institute Management Platform.
  title: Institute Management Backend API
  version: "1.0"
paths:
...
  /institutions:
    get:
      description: Retrieves a paginated list of institutions. Accessible only by
        Super Admin.
      parameters:
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - default: 10
        description: Pagination limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Institutions Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List Institutions
      tags:
      - Institution
    post:
      consumes:
      - application/json
      description: Creates a new institution. Accessible only by Super Admin.
      parameters:
      - description: Institution Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_institution_dto.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Institution Created
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Institution
      tags:
      - Institution
  /institutions/{id}:
    delete:
      description: Soft-deletes an institution by its UUID. Accessible only by Super
        Admin.
      parameters:
      - description: Institution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Institution Deleted
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Institution
      tags:
      - Institution
    get:
      description: Retrieves an institution by its UUID. Accessible only by Super
        Admin.
      parameters:
      - description: Institution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Institution Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Institution
      tags:
      - Institution
    patch:
      consumes:
      - application/json
      description: Partially updates an institution. Accessible only by Super Admin.
      parameters:
      - description: Institution ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_institution_dto.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Institution Updated
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_institution_dto_InstitutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Institution
      tags:
      - Institution
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"